inspectkv: rename package `inspectkv` to `admin`
This package does more than inspectkv, rename to `admin` is better.
