+++ admin/admin_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package inspectkv
+package admin
 
 import (
 	"fmt"
@@ -302,7 +302,7 @@ func (s *testSuite) TestScan(c *C) {
 }
 
 func newDiffRetError(prefix string, ra, rb *RecordData) string {
-	return fmt.Sprintf("[inspectkv:1]%s:%v != record:%v", prefix, ra, rb)
+	return fmt.Sprintf("[admin:1]%s:%v != record:%v", prefix, ra, rb)
 }
 
 func (s *testSuite) testTableData(c *C, tb table.Table, rs []*RecordData) {
@@ -352,7 +352,7 @@ func (s *testSuite) testTableData(c *C, tb table.Table, rs []*RecordData) {
 
 	errRs := append(rs, &RecordData{Handle: int64(1), Values: types.MakeDatums(int64(3))})
 	err = CompareTableRecord(txn, tb, errRs, false)
-	c.Assert(err.Error(), DeepEquals, "[inspectkv:2]handle:1 is repeated in data")
+	c.Assert(err.Error(), DeepEquals, "[admin:2]handle:1 is repeated in data")
 }
 
 func (s *testSuite) testIndex(c *C, tb table.Table, idx table.Index) {
